Extend Integer not Numeric

+++ lib/string_with_unit/core_ext/integer.rb
@@ -1,4 +1,4 @@
-class Numeric
+class Integer
   def to_string_with_unit(singular_unit, plural_unit)
     StringWithUnit::Value.new(self, singular_unit, plural_unit).to_s
   end
